Better way to find max length of the possible mult[].
Added a test case that is useful with the new findValidLength().

Notice a correction where 'r' should be passed as one of the method arguments,
though 'r' can be 'i' and not necessarily 'i+1' as we have seen in most
backtracking examples.

NOTE: Here we are finding all the possible ways we can get the desired
sum. This is not needed as we can quit after the first solution that is
found.