Sanitized the input passwords.
There are two use cases where we can reduce the passwords that can be used
to form various permutations.
1) Consider the case: passwords - {"what", "is", "whatever", "because"},
input - "whateveriswhat"; here the input string does not contain
"because" at all, so there is no point in trying for its permutations.
Thus, the validPasswords are - {"what", "is", "whatever"}.
2) Consider the case: passwords - {"a", "aa", "aaa"},
input - "aaaaab"; here none of the passwords contain the character 'b',
hence there is no point in trying the passwords at all.